Move tracing macro out of partest, closer to user
Ideally, the `tracing` macro would be factored out to a library of testing
macros shipped with partest (as originally done in c619f94a).

Unfortunately, our macro implementation is not stable enough to invoke a macro
that was compiled with an older version, say 2.11.0-M3. More precisely,
compatibility with M3 was broken when the pickling of macro bindings changed in
82f0925b69.

As we want to move partest to its own project and compile it with a released
version of Scala, we must start anticipating this build sequence by building
partest with a released starr&locker while using the quick compiler to compile tests.