SI-7265 General test for spec version
The test for isJavaAtLeast uses the specification.version.

The method argument must have the form "major.minor".

The scaladoc is updated to reflect the new reality and a
test is added under junit.

Note that this implementation aims to be a simple
compromise between the functional and imperative camps,
that is, to be free of both closures and while loops.
And to elicit no cruft like regexes and wrappers for strings.
No doubt even more could be done in this department, but we
don't wish to spoil the fun on codegolf.stackexchange.com.
However, we might decide to sponsor a new site:
codereviewpingpong.com

For 2.10.x, javaSpecVersion is provided as a private member.
The active test is under `run` and the `junit` test must
bide its time in `pending`.

For 2.11, the private members can be public and the app test
replaced with the unit test.