introduces extensibility hooks into the reifier
Quasiquoting macros are surprisingly similar to reifying macros, since
both take something and then produce Scala ASTs for that something.

Therefore the upcoming quasiquote patch reuses the vanilla reifier,
adjusting it in key points to enable splicing and extraction to
support string interpolation syntax.

In this commit we prepare the reifier for being reused later on,
adding a modest amound of extensibility hooks.