extensibility hooks for parser
This is the second of the two patches to the parser necessary for
quasiquotes to function.

This one applies just a couple of minor changes to the way parser works,
so that quasiquotes will be able to extend it to support some
corner cases arising from splicing (see the subsequent quasiquote commit
for more details).