Commit .gitignore directly
Rather than relying on the cloner to copy the
provided gitignore.SAMPLE files.

This finishes the job started in c48509598, mostly
by reverting that commit and moving the two existing
SAMPLE files to the final destinations.

Use `.git/info/exclude` to augment the list of patterns
with entries specific to your workflow.