SI-7689 Fix typing regression with default arguments
Regressed in e28c3edda4. That commit encoded upper/lower bounds
of Any/Nothing as EmptyTree, which were triggering the .isEmpty
check in Namers#TypeTreeSubstitutor and resulting in the default
getter having TypeTree() as the return type. This resulted in a
"missing parmameter type" error.

This commit tightens up that condition to only consider empty
TypeTrees (those wrapping null or NoType.)