Address TODOs around SymbolLoaders and SymbolTable.
SymbolTable refactoring introduced some TODOs that were supposed
to be addressed after M5 release. The reason I couldn't address
those problems right away was a conflict with our plans to modularize
Scaladoc and interactive. However, we decided to delay that work
until after M5 is released so addressing TODOs is not blocked
anymore.

This commit introduces the following changes:

  * Eclipse project definitions for interactive and scaladoc
    depend on scala-compiler project so they are builded against
    latest version of the compiler (quick) instead of STARR.
    This aligns our Eclipse project definitions with build.xml
    structure.
  * Introduce GlobalSymbolLoaders class which wires dependencies
    of SymbolLoaders with assumption of dependency on Global.
  * Switch to GlobalSymbolLoaders in BrowsingLoaders,
    interactive Global and ScaladocGlobal; this eliminates all
    TODO comments introduced before