Simplified elusive typedConstructorPattern.
That was a lot harder than it looks. The dueling test cases
for various corners make it sensitive to the slightest change
in the logic. And of course it never helps that correctness
depends on ad hoc in-place overloading resolution, which is
now performed in a method called inPlaceAdHocOverloadingResolution
so it's a little more obvious.